Ask for consent before starting update checks
NewPipe is contacting its servers without asking for the users' contensent. This is categorized as "tracking" by F-Droid (see https://github.com/TeamNewPipe/NewPipe/discussions/10785).
This commit disabled checking for udpates by default and adds an AlertDialog asking for the user's consent if the app version is eligable for them.